+++ CHANGELOG.md
@@ -12,12 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `cancelOrderWithSigner` function to enable cancel order with l2signer
 - Added `WalletConnection` type
 - Added `transferWithSigner` type
+- Added `batchNftTransferWithSigner` type
 
 ### Deprecated
 
 - `createOrder`, use `createOrderWithSigner` instead
 - `cancelOrder`, use `cancelOrderWithSigner` instead
 - `transfer`, use `transferWithSigner` instead
+- `batchNftTransfer`, use `batchNftTransferWithSigner` instead
 
 ## [0.5.0] - 2022-07-12